move:#36 UpcomingPlans 컴포넌트 schedule 폴더로 이동
